+++ backend/src/hatchling/cli/metadata/__init__.py
@@ -7,6 +7,7 @@ def metadata_impl(called_by_app, field, compact):
 
     from hatchling.bridge.app import get_application
     from hatchling.metadata.core import ProjectMetadata
+    from hatchling.metadata.utils import resolve_metadata_fields
     from hatchling.plugin.manager import PluginManager
 
     app = get_application(called_by_app)
@@ -14,32 +15,18 @@ def metadata_impl(called_by_app, field, compact):
     root = os.getcwd()
     plugin_manager = PluginManager()
     project_metadata = ProjectMetadata(root, plugin_manager)
-    core_metadata = project_metadata.core
 
-    # https://peps.python.org/pep-0621/
-    metadata = {
-        'name': core_metadata.name,
-        'version': project_metadata.version,
-        'description': core_metadata.description,
-        'readme': core_metadata.readme,
-        'requires-python': core_metadata.requires_python,
-        'license': core_metadata.license_expression or core_metadata.license,
-        'authors': core_metadata.authors,
-        'maintainers': core_metadata.maintainers,
-        'keywords': core_metadata.keywords,
-        'classifiers': core_metadata.classifiers,
-        'urls': core_metadata.urls,
-        'scripts': core_metadata.scripts,
-        'gui-scripts': core_metadata.gui_scripts,
-        'entry-points': core_metadata.entry_points,
-        'dependencies': core_metadata.dependencies,
-        'optional-dependencies': core_metadata.optional_dependencies,
-    }
-    if field:
+    metadata = resolve_metadata_fields(project_metadata)
+    if field:  # no cov
         if field not in metadata:
             app.abort(f'Unknown metadata field: {field}')
+        elif field == 'readme':
+            app.display_info(metadata[field]['text'])
+        elif isinstance(metadata[field], str):
+            app.display_info(metadata[field])
+        else:
+            app.display_info(json.dumps(metadata[field], indent=4))
 
-        app.display_info(metadata[field])
         return
 
     for key, value in list(metadata.items()):
@@ -48,7 +35,7 @@ def metadata_impl(called_by_app, field, compact):
 
     if compact:
         app.display_info(json.dumps(metadata, separators=(',', ':')))
-    else:
+    else:  # no cov
         app.display_info(json.dumps(metadata, indent=4))